funcionando con nombres y puntajes, inputs en archivo aparte
