+++ utils.js
@@ -1,0 +1,15 @@
+const readline = require('readline');
+
+const askQuestion = function(query) {
+    const rl = readline.createInterface({
+        input: process.stdin,
+        output: process.stdout,
+    });
+
+    return new Promise(resolve => rl.question(query, ans => {
+        rl.close();
+        resolve(ans);
+    }))
+}
+
+module.exports = {askQuestion}
